+++ src/Helper/WebformHelperSF1601.php
@@ -10,6 +10,7 @@ use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Logger\LoggerChannelFactoryInterface;
 use Drupal\Core\Logger\LoggerChannelInterface;
 use Drupal\os2forms_cpr_lookup\Service\CprServiceInterface;
+use Drupal\os2forms_cvr_lookup\Service\CvrServiceInterface;
 use Drupal\os2forms_digital_post\Exception\InvalidRecipientIdentifierElementException;
 use Drupal\os2forms_digital_post\Exception\RuntimeException;
 use Drupal\os2forms_digital_post\Exception\SubmissionNotFoundException;
@@ -69,6 +70,13 @@ final class WebformHelperSF1601 implements LoggerInterface {
   protected CprServiceInterface $cprService;
 
   /**
+   * The CVR service.
+   *
+   * @var \Drupal\os2forms_cvr_lookup\Service\CvrServiceInterface
+   */
+  protected CvrServiceInterface $cvrService;
+
+  /**
    * The MeMo helper.
    *
    * @var MeMoHelper
@@ -97,6 +105,7 @@ final class WebformHelperSF1601 implements LoggerInterface {
     CertificateLocatorHelper $certificateLocatorHelper,
     EntityTypeManagerInterface $entityTypeManager,
     CprServiceInterface $cprService,
+    CvrServiceInterface $cvrService,
     MeMoHelper $meMoHelper,
     LoggerChannelFactoryInterface $loggerChannelFactory
   ) {
@@ -105,6 +114,7 @@ final class WebformHelperSF1601 implements LoggerInterface {
     $this->webformSubmissionStorage = $entityTypeManager->getStorage('webform_submission');
     $this->queueStorage = $entityTypeManager->getStorage('advancedqueue_queue');
     $this->cprService = $cprService;
+    $this->cvrService = $cvrService;
     $this->meMoHelper = $meMoHelper;
     $this->logger = $loggerChannelFactory->get('os2forms_digital_post');
     $this->submissionLogger = $loggerChannelFactory->get('webform_submission');
@@ -143,8 +153,6 @@ final class WebformHelperSF1601 implements LoggerInterface {
         $message));
     }
 
-    // @todo handle CVR recipient
-    $recipientIdentifierType = 'CPR';
     $recipientIdentifier = $submissionData[$recipientIdentifierKey] ?? NULL;
     if (NULL === $recipientIdentifier) {
       $message = 'Recipient identifier element (key: @element_key) not found in submission';
@@ -157,12 +165,29 @@ final class WebformHelperSF1601 implements LoggerInterface {
         $message));
     }
 
-    try {
-      // Validate recipient identifier.
-      $cprServiceResult = $this->cprService->search($recipientIdentifier);
+    // Remove all non-digits from recipient identifier.
+    $recipientIdentifier = preg_replace('/[^\d]+/', '', $recipientIdentifier);
+
+    $cprServiceResult = NULL;
+    $cvrServiceResult = NULL;
+
+    if (preg_match('/^\d{8}$/', $recipientIdentifier)) {
+      try {
+        $cvrServiceResult = $this->cvrService->search($recipientIdentifier);
+        $recipientIdentifierType = 'CVR';
+      }
+      catch (ServiceException $serviceException) {
+        throw new RuntimeException('Cannot validate recipient CVR');
+      }
     }
-    catch (ServiceException $serviceException) {
-      throw new RuntimeException('Cannot validate recepient identifier');
+    else {
+      try {
+        $cprServiceResult = $this->cprService->search($recipientIdentifier);
+        $recipientIdentifierType = 'CPR';
+      }
+      catch (ServiceException $serviceException) {
+        throw new RuntimeException('Cannot validate recipient CPR');
+      }
     }
 
     $senderSettings = $this->settings->getSender();
@@ -176,7 +201,7 @@ final class WebformHelperSF1601 implements LoggerInterface {
       WebformHandlerSF1601::SENDER_LABEL => $handlerMessageSettings[WebformHandlerSF1601::SENDER_LABEL],
       WebformHandlerSF1601::MESSAGE_HEADER_LABEL => $handlerMessageSettings[WebformHandlerSF1601::MESSAGE_HEADER_LABEL],
     ];
-    $message = $this->meMoHelper->buildMessage($submission, $messageOptions, $handlerSettings, $submissionData, $cprServiceResult);
+    $message = $this->meMoHelper->buildMessage($submission, $messageOptions, $handlerSettings, $submissionData, $cprServiceResult ?? $cvrServiceResult);
 
     $options = [
       'test_mode' => (bool) $this->settings->getTestMode(),