+++ src/Helper/WebformHelperSF1601.php
@@ -10,6 +10,7 @@ use Drupal\Core\Entity\EntityTypeManagerInterface;
 use Drupal\Core\Logger\LoggerChannelFactoryInterface;
 use Drupal\Core\Logger\LoggerChannelInterface;
 use Drupal\os2forms_cpr_lookup\Service\CprServiceInterface;
+use Drupal\os2forms_cvr_lookup\Service\CvrServiceInterface;
 use Drupal\os2forms_digital_post\Exception\InvalidRecipientIdentifierElementException;
 use Drupal\os2forms_digital_post\Exception\RuntimeException;
 use Drupal\os2forms_digital_post\Exception\SubmissionNotFoundException;
@@ -69,11 +70,25 @@ final class WebformHelperSF1601 implements LoggerInterface {
   protected CprServiceInterface $cprService;
 
   /**
+   * The CVR service.
+   *
+   * @var \Drupal\os2forms_cvr_lookup\Service\CvrServiceInterface
+   */
+  protected CvrServiceInterface $cvrService;
+
+  /**
    * The MeMo helper.
    *
    * @var MeMoHelper
    */
   protected MeMoHelper $meMoHelper;
+
+  /**
+   * The Beskedfordeler helper.
+   *
+   * @var BeskedfordelerHelper
+   */
+  private BeskedfordelerHelper $beskedfordelerHelper;
 
   /**
    * The logger.
@@ -97,7 +112,9 @@ final class WebformHelperSF1601 implements LoggerInterface {
     CertificateLocatorHelper $certificateLocatorHelper,
     EntityTypeManagerInterface $entityTypeManager,
     CprServiceInterface $cprService,
+    CvrServiceInterface $cvrService,
     MeMoHelper $meMoHelper,
+    BeskedfordelerHelper $beskedfordelerHelper,
     LoggerChannelFactoryInterface $loggerChannelFactory
   ) {
     $this->settings = $settings;
@@ -105,7 +122,9 @@ final class WebformHelperSF1601 implements LoggerInterface {
     $this->webformSubmissionStorage = $entityTypeManager->getStorage('webform_submission');
     $this->queueStorage = $entityTypeManager->getStorage('advancedqueue_queue');
     $this->cprService = $cprService;
+    $this->cvrService = $cvrService;
     $this->meMoHelper = $meMoHelper;
+    $this->beskedfordelerHelper = $beskedfordelerHelper;
     $this->logger = $loggerChannelFactory->get('os2forms_digital_post');
     $this->submissionLogger = $loggerChannelFactory->get('webform_submission');
   }
@@ -124,10 +143,6 @@ final class WebformHelperSF1601 implements LoggerInterface {
    *   [The response, The MeMo message].
    */
   public function sendDigitalPost(WebformSubmissionInterface $submission, array $handlerSettings, array $submissionData = []): array {
-    $logContext = [
-      'webform_submission' => $submission,
-    ];
-
     $submissionData = $submissionData + $submission->getData();
 
     $handlerMessageSettings = $handlerSettings[WebformHandlerSF1601::MEMO_MESSAGE];
@@ -143,8 +158,6 @@ final class WebformHelperSF1601 implements LoggerInterface {
         $message));
     }
 
-    // @todo handle CVR recipient
-    $recipientIdentifierType = 'CPR';
     $recipientIdentifier = $submissionData[$recipientIdentifierKey] ?? NULL;
     if (NULL === $recipientIdentifier) {
       $message = 'Recipient identifier element (key: @element_key) not found in submission';
@@ -157,12 +170,29 @@ final class WebformHelperSF1601 implements LoggerInterface {
         $message));
     }
 
-    try {
-      // Validate recipient identifier.
-      $cprServiceResult = $this->cprService->search($recipientIdentifier);
+    // Remove all non-digits from recipient identifier.
+    $recipientIdentifier = preg_replace('/[^\d]+/', '', $recipientIdentifier);
+
+    $cprServiceResult = NULL;
+    $cvrServiceResult = NULL;
+
+    if (preg_match('/^\d{8}$/', $recipientIdentifier)) {
+      try {
+        $cvrServiceResult = $this->cvrService->search($recipientIdentifier);
+        $recipientIdentifierType = 'CVR';
+      }
+      catch (ServiceException $serviceException) {
+        throw new RuntimeException('Cannot validate recipient CVR');
+      }
     }
-    catch (ServiceException $serviceException) {
-      throw new RuntimeException('Cannot validate recepient identifier');
+    else {
+      try {
+        $cprServiceResult = $this->cprService->search($recipientIdentifier);
+        $recipientIdentifierType = 'CPR';
+      }
+      catch (ServiceException $serviceException) {
+        throw new RuntimeException('Cannot validate recipient CPR');
+      }
     }
 
     $senderSettings = $this->settings->getSender();
@@ -176,7 +206,7 @@ final class WebformHelperSF1601 implements LoggerInterface {
       WebformHandlerSF1601::SENDER_LABEL => $handlerMessageSettings[WebformHandlerSF1601::SENDER_LABEL],
       WebformHandlerSF1601::MESSAGE_HEADER_LABEL => $handlerMessageSettings[WebformHandlerSF1601::MESSAGE_HEADER_LABEL],
     ];
-    $message = $this->meMoHelper->buildMessage($submission, $messageOptions, $handlerSettings, $submissionData, $cprServiceResult);
+    $message = $this->meMoHelper->buildMessage($submission, $messageOptions, $handlerSettings, $submissionData, $cprServiceResult ?? $cvrServiceResult);
 
     $options = [
       'test_mode' => (bool) $this->settings->getTestMode(),
@@ -188,10 +218,7 @@ final class WebformHelperSF1601 implements LoggerInterface {
     $type = $handlerMessageSettings[WebformHandlerSF1601::TYPE] ?? SF1601::TYPE_DIGITAL_POST;
     $response = $service->kombiPostAfsend($transactionId, $type, $message);
 
-    $this->notice('Digital post sent', $logContext + [
-      'handler_id' => 'os2forms_digital_post',
-      'operation' => 'digital post sent',
-    ]);
+    $this->beskedfordelerHelper->createMessage($submission->id(), $message, (string) $response->getContent());
 
     return [$response, $service->getLastKombiMeMoMessage()];
   }
@@ -300,6 +327,45 @@ final class WebformHelperSF1601 implements LoggerInterface {
 
       return JobResult::failure($e->getMessage());
     }
+  }
+
+  /**
+   * Process Beskedfordeler data.
+   */
+  public function processBeskedfordelerData(int $submissionId, array $data) {
+    $webformSubmission = $this->loadSubmission($submissionId);
+    if (NULL !== $webformSubmission) {
+      $context = [
+        'webform_submission' => $webformSubmission,
+        'handler_id' => 'os2forms_digital_post',
+      ];
+      $status = $data['TransaktionsStatusKode'];
+
+      if (!empty($data['FejlDetaljer'])) {
+        $this->error('@status; @error_code: @error_text', $context + [
+          'operation' => 'digital post failed',
+          '@status' => $status,
+          '@error_code' => $data['FejlDetaljer']['FejlKode'],
+          '@error_text' => $data['FejlDetaljer']['FejlTekst'],
+          'data' => $data,
+        ]);
+      }
+      else {
+        $this->info('@status', $context + [
+          'operation' => 'digital post success',
+          '@status' => $status,
+        ]);
+      }
+    }
+  }
+
+  /**
+   * Proxy for BeskedfordelerHelper::deleteMessages().
+   *
+   * @see BeskedfordelerHelper::deleteMessages()
+   */
+  public function deleteMessages(array $webformSubmissions) {
+    $this->beskedfordelerHelper->deleteMessages($webformSubmissions);
   }
 
 }