+++ CHANGELOG.md
@@ -9,6 +9,13 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.2.0]
+
+### Added
+
+- Added handling of CVR recipients.
+- Added handling of Beskedfordeler messages.
+
 ### Changed
 
 - Update dompdf dependency.
@@ -34,7 +41,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   “KombiPostAfsend”](https://digitaliseringskataloget.dk/integration/sf1601).
 - Added GitHub Actions for coding standards checks and code analysis.
 
-[Unreleased]: https://github.com/itk-dev/os2forms_digital_post/compare/1.1.2...HEAD
+[Unreleased]: https://github.com/itk-dev/os2forms_digital_post/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/itk-dev/os2forms_digital_post/compare/1.1.2...1.2.0
 [1.1.2]: https://github.com/itk-dev/os2forms_digital_post/compare/1.1.1...1.1.2
 [1.1.1]: https://github.com/itk-dev/os2forms_digital_post/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/itk-dev/os2forms_digital_post/compare/1.0.2...1.1.0